requirements.txt
asgiref==3.7.2
boto3==1.28.29
botocore==1.31.29
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.2
django-countries==7.5.1
django-heroku==0.3.1
django-import-export==3.2.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-storages==1.13.2
et-xmlfile==1.1.0
flake8==6.1.0
gunicorn==21.2.0
jmespath==1.0.1
MarkupPy==1.14
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
phonenumbers==8.13.15
Pillow==9.5.0
psycopg2==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.0
urllib3==1.26.16
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
